+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] (0.1.0)
+## Added
+- [dkr sh CONTAINER_NAME](./README.md#sh) command.
+
+## Changed
+- A `--version` flag now prints version of `dkr`, `docker` and `docker-compose`.
+
 ## Fixed
 - Index out of range error when calling `dkr` or `dkr c` without any more arguments.